+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
   - Fix TestName 
+  - Fix direct execution of tests when the script is dot-sourced to global scope
 
 ## 3.0.0-beta (June 24, 2014)
   - Add full support for module mocking