+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
   - Fix TestName 
+  - Output exception source when test fails [GH-147]
 
 ## 2.1.0 (June 15, 2014)
   - Process It blocks in memory [GH-123]