+++ CHANGELOG.md
@@ -5,7 +5,10 @@
   - Remove Validate-Xml function
   - Remove legacy object adaptations support
   - Remove tests testing usage of the global scope
-  - Add function name to Code coverage output
+  - Add function name to Code coverage output [GH-152]
+  - Suppress pipeline output in Context / Describe [GH-155]
+  - Coverage Output Update [GH-156]
+
 
 ## 3.0.0-beta2 (July 4, 2014)
   - Add code coverage [GH-148]