+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
   - Fix TestName 
   - Output exception source when test fails [GH-147]
+  - Add initial implementation of BeforeEach / AfterEach [GH-158]
 
 ## 2.1.0 (June 15, 2014)
   - Process It blocks in memory [GH-123]