+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - Rolling back some earlier Pester Scope changes [GH-164]
   - Legacy expectations cleanup [GH-165]
   - Invoke-Pester tests path fix [GH-166]
+  - Assert-MockCalled default ModuleName fix. [GH-167]
 
 ## 3.0.0-beta2 (July 4, 2014)
   - Add code coverage [GH-148]