+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   - Assert-MockCalled default ModuleName fix. [GH-167]
   - Output exception source when test fails [GH-147]
   - Fix for PesterThrowFailureMessage on PowerShell 2.0. [GH-171]
+  - Pester.bat no longer enables StrictMode.  [GH-172]
 
 ## 3.0.0-beta2 (July 4, 2014)
   - Add code coverage [GH-148]