+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   - Output exception source when test fails [GH-147]
   - Fix for PesterThrowFailureMessage on PowerShell 2.0. [GH-171]
   - Pester.bat no longer enables StrictMode.  [GH-172]
+  - Fixed default behavior of fixture parameter in Describe and Context.  [GH-174]
 
 ## 3.0.0-beta2 (July 4, 2014)
   - Add code coverage [GH-148]