+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   - Fix for PesterThrowFailureMessage on PowerShell 2.0. [GH-171]
   - Pester.bat no longer enables StrictMode.  [GH-172]
   - Fixed default behavior of fixture parameter in Describe and Context.  [GH-174]
+  - Syntax errors in test files, as well as terminating errors from Describe or Context blocks are now treated as failed tests.  [GH-168]
 
 ## 3.0.0-beta2 (July 4, 2014)
   - Add code coverage [GH-148]