+++ CHANGELOG.md
@@ -20,6 +20,7 @@
   - Fixed default behavior of fixture parameter in Describe and Context.  [GH-174]
   - Syntax errors in test files, as well as terminating errors from Describe or Context blocks are now treated as failed tests.  [GH-168]
   - Mock lifetime is no longer tied to It blocks. [GH-176]
+  - Add module manifest
 
 ## 3.0.0-beta2 (July 4, 2014)
   - Add code coverage [GH-148]