+++ CHANGELOG.md
@@ -26,6 +26,7 @@
   - Fixed stack trace information from Describe and Context block errors in PowerShell 2.0. [GH-186]
   - Fixed a problem with parameter / argument resolution in mocked cmdlets / advanced functions.  [GH-187]
   - Improved error reporting when Pester commands are called outside of a Describe block. [GH-188]
+  - Extensive updates to help files and comment-based help for v3.0 release. [GH-190]
 
 ## 3.0.0-beta2 (July 4, 2014)
   - Add code coverage [GH-148]